refactor : 시스템 외부와의 통신과 관련된 S3Service 위치 infrastructure 패키지 밑으로 이동
